+++ pdf/test/all_tests.dart
@@ -40,6 +40,7 @@ import 'widget_form_test.dart' as widget_form;
 import 'widget_grid_view_test.dart' as widget_grid_view;
 import 'widget_multipage_test.dart' as widget_multipage;
 import 'widget_opacity_test.dart' as widget_opacity;
+import 'widget_outline_test.dart' as widget_outline;
 import 'widget_partitions_test.dart' as widget_partitions;
 import 'widget_table_test.dart' as widget_table;
 import 'widget_test.dart' as widget;
@@ -73,6 +74,7 @@ void main() {
   widget_grid_view.main();
   widget_multipage.main();
   widget_opacity.main();
+  widget_outline.main();
   widget_partitions.main();
   widget_table.main();
   widget_text.main();